Finding max using dilation and peak interpolation
Added a function to find the max using the dilation method, with a
window of size 50. Changed the buffer length to 14,700 so that each bin
is 6Hz long. Added kdf constant, which is the size of each bin.
Added a function to interpolate the actual peak. It is possibly not
working. Please double check with slides.